chore: resolve all clippy warnings in examples and tests
Fix ~50 clippy warnings across examples, integration tests, and lib tests:
- Replace approximate PI literal with std::f64::consts::PI
- Fix double-reference clone (use (*k).clone())
- Scope MutexGuard to avoid holding across await points
- Replace loop-variable indexing with allow attribute where intentional
- Use struct initializer syntax instead of field reassign after Default
- Replace redundant pattern matching with .is_pending()
- Remove unused Poll import
- Suppress dead_code for test utility method
- Use drop() instead of let _ = on Futures

Zero clippy warnings remain across lib, tests, and examples.

+++ instancy/src/communication/transport.rs
@@ -618,7 +618,7 @@ mod transport_tests {
         let frame = Frame {
             dataflow_id: DataflowId::from_bytes([1u8; 16]),
             channel_id: 1,
-            payload: vec![0; (MAX_MESSAGE_SIZE as usize) + 1],
+            payload: vec![0; MAX_MESSAGE_SIZE + 1],
         };
         let result = writer.write_frame(&frame).await;
         assert!(result.is_err());
